strukturbaum in main_widget ausgegliedert
